fix(qml): resolve double promotion warning in setAlphaF calls

+++ src/qml_frontend/QmlLineSeries.cpp
@@ -215,7 +215,7 @@ QSGNode* QmlLineSeries::updatePaintNode(QSGNode* oldNode, UpdatePaintNodeData* u
     // QColor 의 알파 채널을 그대로 사용하며, 노드에 Opaque 플래그가 없으므로
     // 알파 합성이 적용된다. 겹치는 시리즈가 예측 가능하게 섞인다.
     QColor color = m_series->color();
-    color.setAlphaF(color.alphaF() * m_series->opacity());
+    color.setAlphaF(static_cast<float>(static_cast<double>(color.alphaF()) * m_series->opacity()));
 
     // 1. 그릴 데이터가 없는 경우
     if (!model || model->pointCount() == 0) {
